controller code

+++ FinalProject/FinalProject/Models/Flight.cs
@@ -20,7 +20,7 @@ namespace FinalProject.Models
         [DataType(DataType.Date)]
         [DisplayFormat(DataFormatString = "{0:yyyy-MM-dd}", ApplyFormatInEditMode = true)]
         public DateTime Date { get; set; }
-        public Int32 BaseFare { get; set; }
+        public Decimal BaseFare { get; set; }
 
         [DataType(DataType.Time)]
         [DisplayFormat(ApplyFormatInEditMode = true, DataFormatString = "{0:HH:mm}")]
